[move scripts] reorganize
